pyproject.toml
[project]
name = "aws-docs-rag-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.1"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.8",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.4.1",
    "typer>=0.24.1",
    "uvicorn>=0.44.0",
]

[project.scripts]
raglab = "app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
